pyproject.toml
[tool.poetry]
name = "conda-config"
version = "0.1.0"
description = "Python library for managing conda configuration"
authors = ["Travis Hathaway <travis.j.hathaway@gmail.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"
packages = [{include = "conda_config"}]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.7"
ruamel-yaml = "^0.17.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
mypy = "^1.2.0"
ipython = "^8.12.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
